conditional build for empty KEY

+++ web/entrypoint.sh
@@ -29,10 +29,12 @@ chmod 644 $BASE_DIR/eum.html
 # apply environment variables to default.conf
 envsubst '${CATALOGUE_HOST} ${USER_HOST} ${CART_HOST} ${SHIPPING_HOST} ${PAYMENT_HOST} ${RATINGS_HOST}' < /etc/nginx/conf.d/default.conf.template > /etc/nginx/conf.d/default.conf
 
-# Patching for Instana tracing
-mv /tmp/ngx_http_opentracing_module.so /usr/lib/nginx/modules
-mv /tmp/libinstana_sensor.so /usr/local/lib
-cat - /etc/nginx/nginx.conf << !EOF! > /tmp/nginx.conf
+if [ -f /tmp/ngx_http_opnetracing_module.so -a -f /tmp/libinstana_sensor.so ]
+then
+    # Patching for Instana tracing
+    mv /tmp/ngx_http_opentracing_module.so /usr/lib/nginx/modules
+    mv /tmp/libinstana_sensor.so /usr/local/lib
+    cat - /etc/nginx/nginx.conf << !EOF! > /tmp/nginx.conf
 # Extra configuration for Instana tracing
 load_module modules/ngx_http_opentracing_module.so;
 
@@ -44,8 +46,12 @@ env INSTANA_MAX_BUFFERED_SPANS;
 env INSTANA_DEV;
 !EOF!
 
-mv /tmp/nginx.conf /etc/nginx/nginx.conf
-echo "{}" > /etc/instana-config.json
+    mv /tmp/nginx.conf /etc/nginx/nginx.conf
+    echo "{}" > /etc/instana-config.json
+else
+    # remove tracing config
+    sed -i '1,3d' /etc/nginx/conf.d/default.conf
+fi
 
 exec nginx-debug -g "daemon off;"
